Add clustering function, calculate centroid etc

+++ src/com/company/Point2D.java
@@ -2,10 +2,10 @@ package com.company;
 
 public class Point2D {
     public int id;
-    public float x;
-    public float y;
+    public double x;
+    public double y;
 
-    public Point2D(int id, float x, float y) {
+    public Point2D(int id, double x, double y) {
         this.id = id;
         this.x = x;
         this.y = y;
@@ -13,8 +13,8 @@ public class Point2D {
 
     public Point2D() {
         this.id = 0;
-        this.x = 0.0f;
-        this.y = 0.0f;
+        this.x = 0.0;
+        this.y = 0.0;
     }
 
 }
